Planarity - somewhat
Adds button and functions to check if the graph is planar or not. This works somewhat. Planarity is checked by determining if the graph has a certain number of edges (as planar graphs have a limit), or a certain number of cycles (as trees are planar, and a graph with a single cycle has a planar embedding).

Implementing a planar testing algorithm is either extremely hard or inefficient -- will come back to this.